+++ packages/wonder-blocks-icon/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-icon
 
+## 5.3.23
+
+### Patch Changes
+
+- Updated dependencies [357a19c]
+- Updated dependencies [6e4158c]
+    - @khanacademy/wonder-blocks-tokens@17.3.0
+
 ## 5.3.22
 
 ### Patch Changes