+++ packages/wonder-blocks-modal/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-modal
 
+## 8.7.11
+
+### Patch Changes
+
+- Updated dependencies [357a19c]
+- Updated dependencies [6e4158c]
+    - @khanacademy/wonder-blocks-tokens@17.3.0
+    - @khanacademy/wonder-blocks-breadcrumbs@3.2.27
+    - @khanacademy/wonder-blocks-icon-button@11.4.5
+    - @khanacademy/wonder-blocks-layout@3.1.60
+    - @khanacademy/wonder-blocks-styles@0.2.53
+    - @khanacademy/wonder-blocks-typography@5.0.3
+    - @khanacademy/wonder-blocks-announcer@1.1.1
+
 ## 8.7.10
 
 ### Patch Changes