+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @khanacademy/wonder-blocks-popover
 
+## 6.3.12
+
+### Patch Changes
+
+- Updated dependencies [357a19c]
+- Updated dependencies [6e4158c]
+    - @khanacademy/wonder-blocks-tokens@17.3.0
+    - @khanacademy/wonder-blocks-icon-button@11.4.5
+    - @khanacademy/wonder-blocks-modal@8.7.11
+    - @khanacademy/wonder-blocks-styles@0.2.53
+    - @khanacademy/wonder-blocks-tooltip@4.2.1
+    - @khanacademy/wonder-blocks-typography@5.0.3
+
 ## 6.3.11
 
 ### Patch Changes